Introduce a basic routing implementation.
This commit introduces a basic routing system. The Router class stores
all routes in a object, grouped by the method. The routes are listed
under each route and regex is used for matching. Once a match is found,
the Router.handle method simply calls the callback provided else throws
a 404 - Not Found Error.